refactor(core): extract createMenuHandlers (PLAN §3.1.f completion)
Pull the 7 menu-lifecycle handlers (openContextMenu,
runPendingDefaultAction, handleContextMenuDismiss,
handleShowContextMenuInstance, handleToggleToolbarMenu,
handleSetDefaultAction, dismissAllPopups) out of init() into a
focused controller.

These handlers were spread across two regions of init() (~lines 482-520
and 1658-1700). They share a tightly-coupled set of dependencies
(toolbarMenu, toolbarStateController, actionContextBuilder, lifecycle,
plugin hooks) so they belong together.

This required moving the actionContextBuilder creation earlier (right
after the prompt-mode handlers and the activationLifecycle) so the
menu-handlers can use buildActionContext. The contextMenuActionContext
memo stays at its original position since it needs the
contextMenuFilePath/ComponentName resources.

The pendingDefaultActionId closure flag is now read+cleared via a
takePendingDefaultActionId callback so the controller can stay pure
about it.

index.tsx: 1952 -> 1881 (-71). 79 context-menu + toolbar specs green.
